requirements.txt
numpy
libcasm-mapping
libcasm-clexmonte
